+++ retail/modules/darkmode.lua
@@ -556,6 +556,8 @@ function BBF.DarkmodeFrames(bypass)
         applySettings(TargetFrame.TargetFrameContainer.BossPortraitFrameTexture, d, v)
         applySettings(FocusFrame.TargetFrameContainer.BossPortraitFrameTexture, d, v)
     end
+    BBF.UpdateClassicEliteOverlay(TargetFrame)
+    BBF.UpdateClassicEliteOverlay(FocusFrame)
 
 
     -- Applying settings based on BetterBlizzFramesDB.darkModeUi value